refactoring...(더는 못 참아..)
