+++ src/components/Layout/Header.js
@@ -3,7 +3,7 @@ import Container from "react-bootstrap/Container";
 import Nav from "react-bootstrap/Nav";
 import Navbar from "react-bootstrap/Navbar";
 import Offcanvas from "react-bootstrap/Offcanvas";
-import { ReactComponent as Logo } from "../assets/Logo.svg";
+import { ReactComponent as Logo } from "../../assets/Logo.svg";
 
 export default function Header() {
   return (
